requirements.txt
discord.py
python-dotenv
requests
psutil
yt-dlp